피드백 반영해 수정
