+++ src/index.js
@@ -1,20 +1,7 @@
 import '../index.css';
-
-// 실제 실행 코드 작성
-// ex)cardGame.start()
-
-//import Card from './Card';
-import {Card, Shuffle} from './Card';
-
-/*const cards = [];
-for (let i = 0; i < 3; i++) {
-  const isGoodCard = (i === 1);
-  const card = new Card(isGoodCard);
-  card.createCardBtn();
-  //cards.push(card);
-  //undercards.appendChild(card.Btn);
-}*/
+import Shuffle from './Shuffle';
 
 const cardCount = 3;
 const shuffle = new Shuffle(cardCount);
+shuffle.shuffleCards();
 shuffle.displayCards();
